+++ Cargo.toml
@@ -31,9 +31,6 @@ hippo-openapi = "0.10"
 hippo = { git = "https://github.com/deislabs/hippo-cli", tag = "v0.16.1" }
 indicatif = "0.17.3"
 is-terminal = "0.4"
-termcolor = "1.2"
-once_cell = "1.0"
-atty = "0.2"
 lazy_static = "1.4.0"
 levenshtein = "1.0.5"
 nix = { version = "0.24", features = ["signal"] }
@@ -50,6 +47,7 @@ semver = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.82"
 sha2 = "0.10.2"
+colors = { path = "crates/colors" }
 spin-app = { path = "crates/app" }
 spin-bindle = { path = "crates/bindle" }
 spin-build = { path = "crates/build" }
