+++ crates/colors/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "colors"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+termcolor = "1.2"
+once_cell = "1.0"
+atty = "0.2"
