+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod build_info;
-pub(crate) mod colors;
 pub mod commands;
 pub mod manifest;
 pub(crate) mod opts;
